+++ src/helper/path.ts
@@ -19,6 +19,8 @@ function getAvatarPath(name: string) {
 
 const ASSETSERVER = 'https://strawberrytree.top/convert/cache'
 
+const RAW_CONTENT_PROXY = "https://strawberrytree.top/raw"
+
 const TranslationIndexUrl = 'https://raw.githubusercontent.com/biuuu/ShinyColors/gh-pages/story.json'
 
 function getAudioPath(id: string, base: string) {
@@ -39,6 +41,16 @@ function getQueryVariable(variable: string) {
             return decodeURIComponent(pair[1]);
         }
     }
+}
+
+function getGithubRawResourcePath(url: string, useProxy=true) {
+    if (url.startsWith("https://github.com")) {
+        url = url.replace("https://github.com", "https://raw.githubusercontent.com").replace("/blob/", "/")
+    }
+    if (useProxy && url.startsWith("https://raw.githubusercontent")) {
+        url = url.replace("https://raw.githubusercontent.com", RAW_CONTENT_PROXY)
+    }
+    return url
 }
 
 // https://github.com/ShinyGroup/SCTranslationData/blob/master/data/story/%E4%B8%89%E5%B3%B0%E7%B5%90%E8%8F%AF/%E3%80%90Hakoni%E2%96%A1a%E3%80%91%E4%B8%89%E5%B3%B0%E7%B5%90%E8%8F%AF/%E3%83%9A%E3%83%AB%E3%82%BD%E3%83%8A%E3%81%AA%E3%82%93%E3%81%A6%E7%AC%91%E3%81%A3%E3%81%A1%E3%82%83%E3%81%86.csv
@@ -190,4 +202,5 @@ export {
     firstJsonUrl,
     queryTranslatedCsv,
     initTranslatedStoryIndex,
+    getGithubRawResourcePath,
 }